style: move CardItem
