+++ src/components/cardItem/CardItem.tsx
